# service/account/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Your API description.",
        "title": "sealos account service",
        "termsOfService": "https://cloud.sealos.io",
        "contact": {
            "email": "bxy4543@gmail.com"
        },
        "version": "v1alpha1"
    },
    "host": "localhost:2333",
    "paths": {
...
        "/admin/v1alpha1/account": {
            "get": {
                "description": "Get user account",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Account"
                ],
                "summary": "Get user account",
                "responses": {
                    "200": {
                        "description": "successfully retrieved user account",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "authenticate error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "failed to get user account",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/admin/v1alpha1/charge": {
            "post": {
                "description": "Charge billing",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Account"
                ],
                "summary": "Charge billing",
                "responses": {
                    "200": {
                        "description": "successfully charged billing",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "authenticate error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "failed to charge billing",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/admin/v1alpha1/real-name-info": {
            "get": {
                "description": "Get user real name info",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Account"
                ],
                "summary": "Get user real name info",
                "responses": {
                    "200": {
                        "description": "successfully retrieved user real name info",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "authenticate error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "failed to get user real name info",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        }
    },
...
}